chore(tools): move constant values, secrets (#4825)
