+++ tools/metabase_dashboard_templates/cli.py
@@ -62,10 +62,7 @@ yaml.safe_dump, then post-processes the text to swap each sentinel
 from __future__ import annotations
 
 import json
-import os
 import re
-import shutil
-import subprocess
 from datetime import datetime
 from pathlib import Path
 from typing import Any, Callable
@@ -76,129 +73,26 @@ import jinja2.meta
 import requests
 import yaml
 
-# ---------------------------------------------------------------------------
-# Environment configuration
-#
-# Maps "staging" and "prod" to the Metabase URL + GCP Secret Manager resource
-# holding that instance's API key.  The interactive wizard uses these.  Each
-# value is overridable at runtime via env var so deployment-specific names
-# don't need a code change:
-#   METABASE_STAGING_URL, METABASE_STAGING_GCP_SECRET
-#   METABASE_PROD_URL,    METABASE_PROD_GCP_SECRET
-# ---------------------------------------------------------------------------
-
-ENVIRONMENTS: dict[str, dict[str, str]] = {
-    "staging": {
-        "url": os.environ.get(
-            "METABASE_STAGING_URL",
-            "https://metabase-staging.dds.dot.ca.gov",
-        ),
-        "gcp_secret": os.environ.get(
-            "METABASE_STAGING_GCP_SECRET",
-            "projects/cal-itp-data-infra-staging/secrets/"
-            "metabase-dashboard-copy-tool-metabase-staging-api-key/versions/latest",
-        ),
-    },
-    "prod": {
-        "url": os.environ.get(
-            "METABASE_PROD_URL",
-            "https://metabase.dds.dot.ca.gov",
-        ),
-        "gcp_secret": os.environ.get(
-            "METABASE_PROD_GCP_SECRET",
-            "projects/cal-itp-data-infra/secrets/"
-            "metabase-dashboard-copy-tool-metabase-prod-api-key/versions/latest",
-        ),
-    },
-}
-
-# ---------------------------------------------------------------------------
-# Strip-key sets
-#
-# Empirically derived from the Metabase OpenAPI spec (v0.58) plus inspection
-# of real GET payloads.  A field is "STRIP" if it is server-generated and
-# either ignored or rejected on write.
-# ---------------------------------------------------------------------------
-
-STRIP_CARD_KEYS = {
-    "id",
-    "entity_id",
-    "created_at",
-    "updated_at",
-    "last_used_at",
-    "view_count",
-    "query_average_duration",
-    "creator_id",
-    "made_public_by_id",
-    "public_uuid",
-    "dashboard_id",
-    "can_write",
-    "card_schema",
-    "metabase_version",
-    "dependency_analysis_version",
-    "archived_directly",
-    "download_perms",
-    "initially_published_at",
-    "cache_invalidated_at",
-    "moderation_reviews",
-    "legacy_query",
-    "is_remote_synced",
-    "source_card_id",
-    "document_id",
-    # Stripped because Metabase recomputes it from dataset_query, and the
-    # source-DB-specific table_id/field_id values inside it are noise.
-    "result_metadata",
-}
-
-STRIP_DASHCARD_KEYS = {
-    "id",
-    "entity_id",
-    "created_at",
-    "updated_at",
-    "dashboard_id",
-    "collection_authority_level",
-    "action_id",
-    # card_id is derived at apply time (each new card gets a fresh id);
-    # virtual dashcards continue to be detected by visualization_settings.virtual_card.
-    "card_id",
-}
-
-STRIP_DASHBOARD_KEYS = {
-    "id",
-    "entity_id",
-    "created_at",
-    "updated_at",
-    "last_viewed_at",
-    "last_used_param_values",
-    "view_count",
-    "creator_id",
-    "made_public_by_id",
-    "public_uuid",
-    "can_write",
-    "can_delete",
-    "can_restore",
-    "archived_directly",
-    "param_fields",
-    "last-edit-info",
-    "collection",
-    "collection_authority_level",
-    "moderation_reviews",
-    "dependency_analysis_version",
-    "initially_published_at",
-    "is_remote_synced",
-}
-
-# Subset of dashboard top-level keys that POST /api/dashboard/ accepts.
-# Anything else gets attached via the follow-up PUT /api/dashboard/{id}.
-POST_DASHBOARD_KEYS = {
-    "name",
-    "description",
-    "collection_id",
-    "collection_position",
-    "parameters",
-    "cache_ttl",
-}
-
+try:  # package context (e.g. `tools.metabase_dashboard_templates.cli` under pytest)
+    from .constants import (
+        POST_DASHBOARD_KEYS,
+        STRIP_CARD_KEYS,
+        STRIP_DASHBOARD_KEYS,
+        STRIP_DASHCARD_KEYS,
+        TEMPLATES_DIR,
+    )
+    from .environments import ENV_LABELS, ENVIRONMENTS
+    from .gcp_secrets import fetch_secret_from_gcp
+except ImportError:  # script context (e.g. `uv run cli.py`, `python cli.py`)
+    from constants import (
+        POST_DASHBOARD_KEYS,
+        STRIP_CARD_KEYS,
+        STRIP_DASHBOARD_KEYS,
+        STRIP_DASHCARD_KEYS,
+        TEMPLATES_DIR,
+    )
+    from environments import ENV_LABELS, ENVIRONMENTS
+    from gcp_secrets import fetch_secret_from_gcp
 
 # ---------------------------------------------------------------------------
 # Metabase HTTP helpers
@@ -256,141 +150,6 @@ def fetch_database_metadata(
         "table_to_id": table_to_id,
         "field_to_id": field_to_id,
     }
-
-
-# ---------------------------------------------------------------------------
-# GCP Secret Manager
-# ---------------------------------------------------------------------------
-
-
-def _run_gcloud_adc_login() -> None:
-    """Shell out to `gcloud auth application-default login`.
-
-    Used when ADC are missing so the user doesn't have to drop out of the
-    wizard to run it themselves.  Inherits stdio so the gcloud browser flow
-    works normally (gcloud prints a URL, the user signs in, control returns).
-    """
-    if shutil.which("gcloud") is None:
-        raise click.ClickException(
-            "gcloud CLI not found on PATH.  Install the Google Cloud SDK "
-            "(https://cloud.google.com/sdk/docs/install) and try again."
-        )
-    click.echo(
-        "Launching `gcloud auth application-default login` "
-        "(a browser tab will open)...",
-        err=True,
-    )
-    try:
-        subprocess.run(["gcloud", "auth", "application-default", "login"], check=True)
-    except subprocess.CalledProcessError as exc:
-        raise click.ClickException(
-            f"`gcloud auth application-default login` failed "
-            f"(exit code {exc.returncode})."
-        )
-    except KeyboardInterrupt:
-        raise click.ClickException("Login cancelled.")
-
-
-def fetch_secret_from_gcp(secret_resource_name: str) -> str:
-    """Fetch a secret payload from GCP Secret Manager via Application Default Credentials.
-
-    `secret_resource_name` must be a fully-qualified resource name:
-        projects/PROJECT_ID/secrets/SECRET_NAME/versions/VERSION
-    Use "latest" as VERSION to always get the current active version.
-
-    If ADC is not configured, this transparently runs
-    `gcloud auth application-default login` and retries once.  All other
-    failures (malformed name, secret missing, permission denied, transport
-    errors) raise ClickException with a message that names the secret.
-    """
-    try:
-        from google.api_core import exceptions as gcp_exceptions
-        from google.auth import exceptions as auth_exceptions
-        from google.cloud import secretmanager
-    except ImportError:
-        raise click.ClickException(
-            "google-cloud-secret-manager is not installed. "
-            "Run: pip install google-cloud-secret-manager"
-        )
-
-    # Sanity-check the resource name format up front, so a typo gives a
-    # clearer error than the server's InvalidArgument.
-    parts = secret_resource_name.split("/")
-    if (
-        len(parts) != 6
-        or parts[0] != "projects"
-        or parts[2] != "secrets"
-        or parts[4] != "versions"
-    ):
-        raise click.ClickException(
-            f"GCP secret resource name is malformed: {secret_resource_name!r}\n"
-            "Expected format: "
-            "projects/PROJECT_ID/secrets/SECRET_NAME/versions/VERSION "
-            "(VERSION may be 'latest')."
-        )
-
-    def _build_client():
-        return secretmanager.SecretManagerServiceClient()
-
-    try:
-        client = _build_client()
-    except auth_exceptions.DefaultCredentialsError:
-        click.echo(
-            "GCP Application Default Credentials are not configured.",
-            err=True,
-        )
-        _run_gcloud_adc_login()
-        # Retry once after login completes.  If it still fails here, the
-        # exception propagates -- something deeper is wrong (e.g. login
-        # succeeded but the credential file is unreadable).
-        client = _build_client()
-
-    try:
-        response = client.access_secret_version(name=secret_resource_name)
-    except auth_exceptions.RefreshError as exc:
-        raise click.ClickException(
-            f"GCP credentials are expired or invalid "
-            f"(needed to fetch {secret_resource_name!r}): {exc}\n"
-            "Refresh and retry.  For cal-itp workforce identity:\n"
-            "  glogin-calitp-staging   # or glogin-calitp-prod\n"
-            "Or run `gcloud auth application-default login` directly."
-        )
-    except gcp_exceptions.Unauthenticated as exc:
-        raise click.ClickException(
-            f"GCP rejected your credentials on {secret_resource_name!r}: {exc}\n"
-            "Refresh ADC and retry: see your workforce identity login alias "
-            "(e.g. glogin-calitp-staging)."
-        )
-    except gcp_exceptions.NotFound:
-        # Reaching this branch is rare -- GCP usually conflates NotFound and
-        # PermissionDenied into the latter to avoid leaking secret names.
-        raise click.ClickException(
-            f"GCP secret not found: {secret_resource_name!r}\n"
-            "Verify with:\n"
-            f"  gcloud secrets describe {parts[3]} --project={parts[1]}"
-        )
-    except gcp_exceptions.PermissionDenied as exc:
-        raise click.ClickException(
-            f"GCP returned 403 on {secret_resource_name!r}: {exc}\n"
-            "This means EITHER the secret does not exist OR your account "
-            "lacks secretmanager.versions.access.  GCP intentionally collapses "
-            "both into the same error.  To distinguish:\n"
-            f"  1. Check existence:\n"
-            f"       gcloud secrets describe {parts[3]} --project={parts[1]}\n"
-            "  2. If `describe` succeeds, you need the role.  Run:\n"
-            f"       gcloud secrets add-iam-policy-binding {parts[3]} \\\n"
-            f"         --project={parts[1]} \\\n"
-            '         --member="user:$(gcloud config get-value account)" \\\n'
-            '         --role="roles/secretmanager.secretAccessor"\n'
-            "  3. If `describe` returns NOT_FOUND, the secret hasn't been "
-            "created yet -- create it (or ask whoever owns the project to)."
-        )
-    except gcp_exceptions.GoogleAPICallError as exc:
-        raise click.ClickException(
-            f"GCP error fetching secret {secret_resource_name!r}: {exc}"
-        )
-
-    return response.payload.data.decode("utf-8").strip()
 
 
 # ---------------------------------------------------------------------------
@@ -1843,17 +1602,6 @@ def cmd_template_to_dashboard(
 # ---------------------------------------------------------------------------
 
 
-ENV_LABELS: dict[str, str] = {
-    "staging": "Metabase Staging",
-    "prod": "Metabase Prod",
-}
-
-# Directory where exported templates live (and where the wizard looks for
-# already-exported ones).  Sits next to this script so the templates ship
-# alongside the tool in source control.
-TEMPLATES_DIR: Path = Path(__file__).resolve().parent / "templates"
-
-
 def _ensure_templates_dir() -> Path:
     """Make sure TEMPLATES_DIR exists; create it (with a .gitkeep) if not.
 
@@ -2061,8 +1809,7 @@ def cmd_interactive(ctx: click.Context) -> None:
 
     Reads Metabase API keys for each instance from GCP Secret Manager via
     Application Default Credentials.  Configure the URL + secret resource
-    name per environment in ENVIRONMENTS at the top of this file (or via the
-    METABASE_{STAGING,PROD}_{URL,GCP_SECRET} env vars).
+    name per environment in ENVIRONMENTS in environments.py.
     """
     templates_dir = _ensure_templates_dir()
     existing_templates = _list_template_files()
